requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
feedparser>=6.0.0
apscheduler>=3.10.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
pytest>=8.0.0
anthropic>=0.34.0
openai>=1.40.0
google-genai>=1.0.0
